+++ flake.nix
@@ -36,7 +36,7 @@
 
         packages.default = pkgs.rustPlatform.buildRustPackage {
           pname = "ngit-grasp";
-          version = "0.1.0";
+          version = "1.0.2";
           src = ./.;
           cargoLock = {
             lockFile = ./Cargo.lock;